+++ src/utils/polyfill.js
@@ -1,10 +1,10 @@
-import { FFT, RFFT } from 'fftw-js'
+import { FFT, RFFT } from './FFTW'
 import CryptoJS from 'crypto-js'
 
-Float32Array.prototype.fft = function () {
+Float64Array.prototype.fft = function () {
   let count = this.length
   const fft = new FFT(count)
-  const transform = Float32Array.from(fft.forward(this))
+  const transform = Float64Array.from(fft.forward(this))
   fft.dispose()
   return transform
 }
@@ -20,23 +20,23 @@ const scaleTransform = function (trans, size) {
   return x;
 }
 
-Float32Array.prototype.ifft = function (length) {
+Float64Array.prototype.ifft = function (length) {
   const fft = new FFT(length)
-  let temp = Float32Array.from(this)
+  let temp = Float64Array.from(this)
   temp = scaleTransform(temp, length)
   const transform = fft.inverse(temp.slice(0, length)).slice(0, length)
   fft.dispose()
   return transform
 }
 
-Float32Array.prototype.fftComplexMultiply = function (other) {
+Float64Array.prototype.fftComplexMultiply = function (other) {
   let length = this.length
   // normalize to even length
   if (length % 2 === 1) {
     length += 1
   }
   // result only store the real parts of complex multiply
-  const result = new Float32Array(length)
+  const result = new Float64Array(length)
   const half = length / 2
   for (let idx = 0; idx < half; ++idx) {
     const i = 2 * idx
@@ -57,9 +57,9 @@ Float32Array.prototype.fftComplexMultiply = function (other) {
   return result
 }
 
-Float32Array.prototype.complexMultiply = function (other) {
+Float64Array.prototype.complexMultiply = function (other) {
   const length = this.length
-  const result = new Float32Array(length)
+  const result = new Float64Array(length)
   for (let idx = 0; idx < length - 1; idx = idx + 2) {
     const a = this[idx]
     const b = this[idx + 1]
@@ -74,22 +74,22 @@ Float32Array.prototype.complexMultiply = function (other) {
   return result
 }
 
-Float32Array.prototype.real = function () {
+Float64Array.prototype.real = function () {
   const count = this.length
   const result = this.filter((value, idx) => idx % 2 === 0 && idx < count - 1)
-  let half = Float32Array.from(result)
+  let half = Float64Array.from(result)
   half = half.reverse().slice(0, count - result.length)
 
   return result.concat(half)
 }
 
-Float32Array.prototype.image = function () {
+Float64Array.prototype.image = function () {
   return this.filter((value, idx) => idx % 2 === 1)
 }
 
-Float32Array.prototype.concat = function (second) {
+Float64Array.prototype.concat = function (second) {
   const firstLength = this.length
-  const result = new Float32Array(firstLength + second.length)
+  const result = new Float64Array(firstLength + second.length)
 
   result.set(this)
   result.set(second, firstLength)
@@ -97,35 +97,35 @@ Float32Array.prototype.concat = function (second) {
   return result
 }
 
-Float32Array.prototype.round = function () {
+Float64Array.prototype.round = function () {
   return this.map(value => Math.round(value))
 }
 
-Float32Array.prototype.mod = function (mod) {
+Float64Array.prototype.mod = function (mod) {
   return this.map(value => value % mod)
 }
 
-Float32Array.prototype.nonzero = function () {
+Float64Array.prototype.nonzero = function () {
   return this.findIndex(value => value != 0)
 }
 
-Float32Array.prototype.zeros = function () {
+Float64Array.prototype.zeros = function () {
   return this.findIndex(value => value == 0)
 }
 
-Float32Array.prototype.multiply = function (number) {
+Float64Array.prototype.multiply = function (number) {
   return this.map(value => value * number)
 }
 
-Float32Array.prototype.clone = function () {
-  return Float32Array.from(this)
+Float64Array.prototype.clone = function () {
+  return Float64Array.from(this)
 }
 
-Float32Array.prototype.toArray = function () {
+Float64Array.prototype.toArray = function () {
   return Array.prototype.slice.call(this)
 }
 
-Float32Array.prototype.toPrintString = function () {
+Float64Array.prototype.toPrintString = function () {
   let result = '['
   const count = this.length
   for (let idx = 0; idx < count; ++idx) {
